% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.clus.list}
\alias{aws.clus.list}
\title{Lists Clusters}
\usage{
aws.clus.list(active = TRUE, since = NULL, before = NULL)
}
\arguments{
\item{active}{if TRUE reports only active clusters}

\item{since}{a date string (that can be converted using as.POSIXct) used for --created-since}

\item{before}{a date string (that can be converted using as.POSIXct) used for --created-before}
}
\value{
a JSON blob of active clusters
}
\description{
Lists Clusters
}

