% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.get}
\alias{aws.get}
\title{get a file to the cluster}
\usage{
aws.get(cl, src, dest = NULL)
}
\arguments{
\item{cl}{is the cluster object returned from \code{aws.clus.create} and friends}

\item{src}{src file/path}

\item{dest}{(defaults to $HOME)}
}
\description{
get a file to the cluster
}
\details{
will send a file from your local end to the cluster
}

