% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.init}
\alias{aws.init}
\title{Initialize the AWS System}
\usage{
aws.init(ec2key = NULL, localpubkey = NULL, optsmore = NULL)
}
\arguments{
\item{ec2key}{this is your EC2 key that you created in the EC2 AWS Console}

\item{localpubkey}{this is for example the contents of
~/.ssh/id_dsa.pub and if provided makes ssh'ing into the cluster
easier}

\item{optsmore}{more options that override the options}
}
\description{
Initialize the AWS System
}
\details{
\code{optsmore} can be used to override the values in
aws.options(). If you don't provide the EC2 key, this
package will find the first available key
}
\examples{
\dontrun{
aws.init(localpub="~/.ssh/id_dsa.pub")
aws.init(localpub="~/.ssh/id_dsa.pub",optsmore=list(customscript='https://raw.githubusercontent.com/saptarshiguha/mozaws/master/bootscriptsAndR/sample.sh'))
}
}

