% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.kill}
\alias{aws.kill}
\title{Kills/Terminates the cluster}
\usage{
aws.kill(clusters)
}
\arguments{
\item{cluster}{object (from \code{aws.clus.create}, \code{aws.clus.info})}
}
\description{
Kills/Terminates the cluster
}

