% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.modify.groups}
\alias{aws.modify.groups}
\title{Add new nodes to a cluster}
\usage{
aws.modify.groups(cl, n, groupid = NULL,
  type = as.character(aws.options()$inst.type["worker"]), spotPrice = NULL,
  name = NULL)
}
\arguments{
\item{cl}{is the cluster object. How many times must i repeat myself?}

\item{n}{number of nodes to add}

\item{groupid}{use this if you need to modify (resize) and exisiting group. These ids can be found via \code{aws.list.groups}}

\item{type}{Instance type for the worker}

\item{spotPrice}{is the price in dollars (numeric) you're willing
to pay. It will choose a default if not given. If \code{spotPrice}
is "ondemand", then OnDemand instances will be started at the OnDemand price}
}
\description{
Add new nodes to a cluster
}
\details{
This returns the info from \code{aws.clus.info}. If you'd
like to delete this group, call the function with the
\code{groupid} and \code{n} set to 0.
}

