% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.send}
\alias{aws.send}
\title{send a file to the cluster}
\usage{
aws.send(cl, src, dst = NULL)
}
\arguments{
\item{cl}{is the cluster object returned from \code{aws.clus.create} and friends}

\item{src}{src file/path}

\item{dst}{(defaults to $HOME)}
}
\description{
send a file to the cluster
}
\details{
will send a file from your local end to the cluster
}

