% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aws.spot.price}
\alias{aws.spot.price}
\title{Get Spot Prices}
\usage{

  aws.spot.price(type = as.character(options("mzaws")[[1]]$inst.type["worker"]),
  hrsInPast = 6)
}
\arguments{
\item{type}{the type of the worker}

\item{hrsInPast}{prices since when as of now}
}
\value{
a data table with prices
}
\description{
Get Spot Prices
}

