% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aws.step.run}
\alias{aws.step.run}
\title{Run a step}
\usage{
aws.step.run(cl, script, wait = TRUE)
}
\arguments{
\item{cl}{is a cluster object returned by \code{aws.clus.create} and friends}

\item{script}{is a URL (not a file name!, something like http://) to download and run. E.g. an Rscript file}
}
\description{
Run a step
}

