% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{makeProgressString}
\alias{makeProgressString}
\title{creates string for tracking progress of the job}
\usage{
makeProgressString(appid, remotenode, port = 4040)
}
\arguments{
\item{appid}{is an application id taken from \code{scApplicationList}. It can be missing anf if is so, it will be taken from the first currently running application}

\item{remotenode}{see \code{scApplicationList}}

\item{port}{is 4040}
}
\description{
creates string for tracking progress of the job
}

