% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{monitorCurrentSparkApplication}
\alias{monitorCurrentSparkApplication}
\title{Will always run and update with the currently running job}
\usage{
monitorCurrentSparkApplication(cl, port = 4040, mon.sec = 5)
}
\arguments{
\item{cl}{is the output from \code{aws.clus.create}}

\item{port}{is 4040}

\item{mon.sec}{is the update frequency}
}
\description{
Will always run and update with the currently running job
}

