% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracker.R
\name{scApplicationList}
\alias{scApplicationList}
\title{Get all applications that talk to spark}
\usage{
scApplicationList(remotenode, port = 4040, verbose = FALSE)
}
\arguments{
\item{remotenode}{a string that is used to talk to the spark server like a ssh command string. See details}

\item{port}{default is 4040 (current running application) or the history server 18080}

\item{verbose}{default is FALSE, when TRUE prints out the main curl command}
}
\description{
Get all applications that talk to spark
}
\details{
Typically the remotenode command is a ssh command which will run a curl query on the remote. See examples.
If \code{remotenode} is missing, it will be taken from the last value in  \code{options('mozremote')}.
See \href{http://spark.apache.org/docs/latest/monitoring.html} for more details
}
\examples{
\dontrun{
cl <- aws.clus.create(workers=2,spark=TRUE,ver=TRUE)
scApplicationList(remote=sprintf("ssh hadoop ATSIGN \%s",cl$MasterPublicDnsName))
scApplicationList() ## takes remote from options("mozremote") which gets populated by aws.clus.create
scApplicationList(remotenode="sh -c") ## If you wish to run this on the remote AWS console, then call
}
}

