% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tracker.R
\name{scJobsForApplication}
\alias{scJobsForApplication}
\title{Get a list of all jobs that were run on the application}
\usage{
scJobsForApplication(appid, remotenode, port = 4040, verbose = FALSE)
}
\arguments{
\item{appid}{is an application id taken from \code{scApplicationList}. It can be missing anf if is so, it will be taken from the first currently running application}

\item{remotenode}{see \code{scApplicationList}.}

\item{port}{is 4040}

\item{verbose}{set to TRUE for actual command run}
}
\description{
Get a list of all jobs that were run on the application
}

