% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{scStages}
\alias{scStages}
\title{Get all stages run on this application}
\usage{
scStages(appid, remotenode, port = 4040, verbose = FALSE)
}
\arguments{
\item{appid}{is an application id taken from \code{scApplicationList}. It can be missing anf if is so, it will be taken from the first currently running application}

\item{remotenode}{see \code{scApplicationList}}

\item{port}{is 4040}

\item{verbose}{set to TRUE for the command}
}
\description{
Get all stages run on this application
}

