% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.awsCluster}
\alias{as.awsCluster}
\title{Converts a cluster-id string into a cluster object}
\usage{
as.awsCluster(clusterid, name = NA)
}
\arguments{
\item{clusterid}{is the id string}

\item{name}{some name you want to give}
}
\description{
Converts a cluster-id string into a cluster object
}

