% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aws.clus.create}
\alias{aws.clus.create}
\title{Create a cluster}
\usage{
aws.clus.create(name = NULL, workers = NULL, master = NULL,
  hadoopops = NULL, timeout = NULL, verbose = FALSE, emrfs = FALSE,
  customscript = NULL, wait = FALSE, spark = FALSE, noR = NULL)
}
\arguments{
\item{name}{is the name of the cluster, if not provided one will be created for you}

\item{workers}{defines the workers, see details}

\item{master}{defines master , see details}

\item{hadoopops}{options that overide 'hadoopops' from aws.options()}

\item{timeout}{over timeout from the options (minutes)}

\item{verbose}{be catty?}

\item{emrfs}{turns on emrfs and consistency}

\item{customscript}{override options}

\item{wait}{TRUE or FALSE for waiting. If FALSE, the function returns immediately or waits}

\item{spark}{if TRUE will start a spark interactive cluster. If a spark cluster is used, you cannot write interactive hadoop jobs.}

\item{noR}{do not install a zillion R packages (if you're using spark-python, you might want to set this to TRUE(which mozaws automatically does anyways) and save time)}
}
\description{
Create a cluster
}
\details{
The arguments \code{hadoopops, timeout, customscript} can
also be set in options. If \code{wait} is FALSE, the function will
return immediately and can be monitored using
\code{aws.clus.wait}. If \code{workers} is a number, then the type
is taken from aws.options(). If a string, this corresponds
to the instance type and the number is taken from
aws.options()$numworkers. If a list, it needs to of the
form \code{list(number, type)}. For \code{master}, it is enough to
leave as NULL (and will be inferred from options) or you pass a
type. The \code{timeout} is a set number of hours after which the
cluster is killed. You'll thank me later.
}
\examples{
\dontrun{
s <- aws.clus.create(workers=1,wait=TRUE,customscript='https://raw.githubusercontent.com/saptarshiguha/mozaws/master/bootscriptsAndR/sample.sh')
s <- aws.clus.create(workers=1)
s <- aws.clus.wait(s)
}
}

