% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.clus.info}
\alias{aws.clus.info}
\title{Describes the cluster}
\usage{
aws.clus.info(cl)
}
\arguments{
\item{cl}{is what is returned from \code{aws.clus.create} or \code{aws.clus.wait}}
}
\value{
an object of awsCluster. Very detailed object. Save it.
}
\description{
Describes the cluster
}

