% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.r
\name{aws.list.groups}
\alias{aws.list.groups}
\title{List the Instance group nodes (Task and Spot nodes)}
\usage{
aws.list.groups(cl, reqGt0 = TRUE)
}
\arguments{
\item{cl}{is the once again the object from \code{aws.clus.create}}
}
\description{
List the Instance group nodes (Task and Spot nodes)
}

