% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aws.r
\name{secondsToString}
\alias{secondsToString}
\title{Takes seconds and converts to human readable format}
\usage{
secondsToString(secs, rnd = 2)
}
\arguments{
\item{secs}{is the number of seconds}
}
\value{
a string
}
\description{
Takes seconds and converts to human readable format
}

